Added basic Accounting and Sheet Plugin.
Still WIP, added the basic Accounting and Sheet logic. Playername parsing is still missing, so it does not yet work.
Also added Embed Plugin, even though this will probably be refactored in the future for more modularity.
